Refactor and correct support for top directories
Deriving top directory to use and failing is now properly implemented.
Now, in case top directory cannot be used, the user is directly provided
an error message. Home trash is not used as a fallback, as a design
decision was taken to not support trashing files from external mounts to
home trash in ANY use case.

That required refactoring the error handling part in the main function.
Will have to refactor to probably use custom error types since some
errors are not supposed to reach the user in specific cases. Need to
double check if all the errors bubbling up are ones that need user's
eyes or if the spec says to fallback on something.