Add directorysizes initial implementation
This is still untested.

Fixed looking up Cargo details during runtime. Cargo env vars are only
available during build time, so had to use the env! macro

Use OpenOptions to atomically create/update trashinfo entry. O_EXCL is
in use when create_new() is added.

Checking if the directory is writable is now done by checking all
permissions, not just W_OK. This is to adhere to spec guidelines on
checking additional permissions if the tool decides.

Added a new function to check if a given file can be deleted by the
current user. This is done by checking the parent directory permissions,
and the actual file permissions.

get_dir_size() is not implemented at the moment. This is a simple
recursive iteration of files to sum up the total size.