Implement dir size calculation
The exact size is not equal to the output from du -B1. There's a few
megs or tens of megs of a difference. However I noticed the same
difference between du -B1 and (ex:) estimation out of Gnome Files.

When testing this I noticed that Gnome Files restoration doesn't even
use the directorysizes file. I can't find why this wasn't implemented,
and I've asked from the community.

If this isn't being used by Gnome, might as well disable with a feature
flag later.