Add dirsizes cleanup when trashing
This takes care of an existence of a trash implementation that doesn't
use directorysizes file that would restore a directory from the trash
and wouldn't update the directorysizes entry.

The trashinfo entry is also fixed to use the urlencoded file name for
the Path key now. This was a bug that wasn't detected earlier.