Add a subtitle to the selection
The subtitle/tooltip? is based on the current sort type. Supposed to be
an easier place to see info the user is interested in rather than the
side window at a glance.

Fixed a bug in writing and reading trashinfo deletion date. Timezone
info should be not in the timestamp, and should be assumed based on the
local timezone.