Fix a few symlink related bugs
Some of the checks and syscalls were traversing symlinks which was not
ideal for a trashing use case. A link should be treated as valid even
when the target doesn't exist.

Also fixed must_have_dir to have mkdir -p behavior instead of mkdir.

Another bug found in deriving the timezone